requirements.txt
google-adk
mindsdb-sdk==3.4.5
vertexai==1.43.0
email-validator==2.2.0
psycopg2-binary
fastapi
uvicorn
PyYAML==6.0.2
google-cloud-storage==2.18.2
numpy==1.26.4
google-cloud-pubsub==2.32.0
croniter==6.0.0
aiohttp==3.13.2
matplotlib
databricks-sql-connector
pandas
openpyxl
google-cloud-bigquery